[ios] Add title and message in the confirmation dialog
This CL adds a title and a message to ActionSheetCoordinator and
specifies the direction of the popover. The dialog always shows up
on the top or bottom of a tab cell on iPad.

Note that the message is different depending on whether a user is
signed in.

Confirmation on the first row on iPad: https://drive.google.com/file/d/1adxqLiT1aLVYjcniX_WjmG_NADtI-P_l/view?usp=sharing
Confirmation on the second row on iPad: https://drive.google.com/file/d/1ibNOuaeoJbS5028bbCoxJnK3xu28nisp/view?usp=sharing

Bug: 329627336
Change-Id: I32b21a38954347eb25badd76f25020134e93d2fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5717729
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329536}