[ios] Move overscroll action image sizeToFit after adding to superview
According to documentation, sizeToFit called on a view that has not been
added to the view hierarchy may resize to the bounds of the screen. This
may be require a alot of memory to size to such a large bounds and
then convert back to the desired size.

Low-Coverage-Reason: EXPERIMENTAL_CODE
Change-Id: Ib2f9e7a73e58b773fdd9ed0f03cac006ea52428d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5719128
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1329565}