ozone: Distinguish BT.601/709 YUV encoding on Vulkan CrOS devices
This CL updates the hw overlay path to be able to treat YUV 4:2:0 BT.601
and BT.709 resources differently, i.e., we plumb the right YUV encoding
to KMS.

The reason this is necessary is that we essentially need to revert
CL:5599625 to be able to fix webgpu CTS tests (e.g., b:345627621).
However, reverting just that CL would sometimes cause a visual
difference between hw overlays and compositing (see b:343793659).
Therefore, in addition to reverting those changes, we need to make sure
that the hw overlay path is consistent with the compositing path.

The reason to limit this CL to Vulkan devices is that I'm not confident
that all ChromeOS devices can use the KMS YUV encoding property
correctly. Instead, I reasoned that devices that are modern enough to
migrate to Vulkan can probably also deal with this property correctly.
See CL:2336347 for historical context on this.

Bug: b:233667677,b:345627621
Test: on volteer - webgpu:web_platform,copyToTexture,video:copy_from_video:videoName=\"four-colors-vp9-bt709.webm\";*
Test: on volteer - webgpu:web_platform,external_texture,video:importExternalTexture*
Test: on volteer - manually play four-colors-vp9-bt709.webm and four-colors-vp9-bt601.webm
Change-Id: Ibee8805c516ebae640e38e58cbf508a262628c46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5703831
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Gil Dekel <gildekel@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329575}