[autofill][clank][metrics] Add Save Card Loading Metrics
Add a metric for showing the loading state and the loading result
after saving a card. The value of the loading shown metrics
("Autofill.CreditCardUpload.LoadingShown") is a boolean and the value
of the result metric ("Autofill.CreditCardUpload.LoadingResult") is an
enum ("AutofillSavePaymentMethodPromptResultEnum") that tracks whether
the bottom sheet was manually closed by the user during loading or
closed automatically after the loading completed.

The save card prompt result enum is used to be consistent with the
enums used for this metric on other platforms (desktop/ios).

Added an additional change to the model used in the unit test from
mock to the real object. The real object is simple to create/use
and simplifies the test (saves us from mocking out reactive behaviour).

Bug: 1499264,b/348375306
Change-Id: Ia260cf22943bf2e0c04f3dac6f5f9f92bf889dd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5717859
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Darwin Yang <darwinyang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329581}