android: Use top resumed position callback for InputHint
Perform InputHintChecker.setView() in the
Activity.onTopResumedActivityChanged() callback. It is called when
activity gets or loses the top resumed position in the system.

This should allow the Input Hint to work in multi-window mode.

Since the InputHintChecker depends on native, the
onTopResumedActivityChangedWithNative() is introduced, analogous to
onResumeWithNative().

Bug: 350531728
Change-Id: Iecaadf1f43830bde26fb9bee053c348ccc838857
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5713149
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329597}