media/gpu/H264VaapiVED: Correct H264 min and max QP values
The values referenced for the original min and max QP values
are criteria of good/bad quality to be used by quality scaler. The
min QP value is 24 is actually too high.
The webrtc encoder is using 0-51. This CL sets min and max QP to 1
and 51, respectively, because the QP=0 is lossless encoding and
it will not be used in Chrome today.

BUG=b:351989151
TEST=video.EncodeAccelp.h264* on brya

Change-Id: I272e7eb566aeb7c964a32d837d44f017a4f95cee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5711490
Auto-Submit: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Justin Green <greenjustin@google.com>
Commit-Queue: Justin Green <greenjustin@google.com>
Cr-Commit-Position: refs/heads/main@{#1329599}