Reapply "Use new v8::Exception::CaptureStackTrace() API for DOMException stacks"
This allows us to put the stack directly on the DOMException's v8
wrapper object, instead of attaching a native data property with a
dummy v8::Object to capture and retrieve the stack. It also
produces a more descriptive error message.

This reverts commit 30cfedd53b5aa4120ce346bf18f603139407f752. The
issue that led to that revert was fixed in
https://chromium-review.googlesource.com/c/v8/v8/+/5690554

Change-Id: I0ce4c3b628ada1e08b5f06dfa8058f37392f6a7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5717853
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Cammie Smith Barnes <cammie@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329604}