[lensoverlay] Calculate an initial wiggle value in constructor.
This makes sure the circles are not overlapping in the same
location. If all of the circles are in the same location, you
are able to see a "halo" or the borders of our gradient color
stops. This also disables wiggling when entering into a region
selection state from invocation.

Change-Id: Ifb0a16ec108c4e42fee0f124f1dae2436fce1b0b
Bug: b:353776428
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5717544
Auto-Submit: Juan Mojica <juanmojica@google.com>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Reviewed-by: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1329612}