Change pointers to refs in TemplateURLService test constructor
We modify the test constructor to take references to PrefService
and SearchEngineChoiceService to make sure that they are never
passed as nullptr.

This is part of a multi-CL change. [3/4]

Bug: 40287734
Change-Id: I44b76663d5e1a297fe36f6c115b0dc6a5227248a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5707969
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Auto-Submit: Jack Yammine <jyammine@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: David Roger <droger@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329644}