iwa: Add ws://proxy_url to injected CSP
This adds the ws:// equivalent of the dev server proxy URL to the
connect-src of the injected CSP when loading an IWA that uses an
http dev proxy URL. Https proxy URLs don't need this as wss connections
are already allowed with the default CSP.

This makes it easier to use standard dev environments which often use
Web Sockets for HMR over an http connection.

This also updates the dev proxy URL validation to enforce that the URL
is potentially trustworthy (i.e. ~SecureContext). Http localhost URLs
will still be allowed, but this will now block remote http URLs.

Bug: 258829385
Change-Id: Iac66aecec7c35036f962c330e89f7d8515f3e477
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5710219
Reviewed-by: Zelin Liu <zelin@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329672}