Only compute deadline in ExternalBeginFrameSourceAndroid in one place
While attempting to mimic some logic from ExtrenalBeginFrameSourceWin for Android
that runs at half frame rate:

  https://source.chromium.org/chromium/chromium/src/+/main:components/viz/service/frame_sinks/external_begin_frame_source_win.cc;l=62;drc=1bf071a86da81bce421efad0a746aabb371888cb

I missed that the Windows version was modifying the deadline based on the doubled
interval, so some of the time BeginFrame calls were happening too early, resulting
in videos dropping frames.

To better align the Android version with the Windows version and reduced the
duplication of the deadline calculation, this change moves the computation in to
OnVsyncImpl instead of the callers.

Change-Id: I2ed50bc506ee4032db8978274c105e44f8f6f4cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5718290
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Bo Liu <boliu@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329701}