[iOS] Show/hide autofill form button based on the focus field type
Before this CL, the autofill form button would be visible on all of the manual fill segments for all of the cells.

This CL changes this behavior such that the autofill form button is visible on the basis of the focused field's type. So, for example, if the focused field is of address type, the autofill form button would be visible in all the cells of the address segment.

`FormInputInteractionDelegate` is added in this CL. It's purpose is to make manual fill classes aware of the changes due to the form interactions by the user. Currently, it informs it's delegates of the focus changes to a new field.

`FormInputAccessoryMediator` holds a weak reference to `FormInputInteractionDelegate` and `ExpandedManualFillCoordinator` implements `FormInputInteractionDelegate`. `FormInputAccessoryCoordinator` is responsible for assigning the `FormInputAccessoryMediator`'s delegate to  `ExpandedManualFillCoordinator`.

On focus changes to new fields, `ExpandedManualFillCoordinator` computes whether the view needs to be updated. There are 2 scenarios where that can happen:
1. Current segment is of type A and focus field changes from type A -> B: In this scenario, the autofill form button needs to be removed.
2. Current segment is of type B and focus field changes from type B -> A. In this scenario, the autofill form button needs to be added to the view.

A parameter `showAutofillFormButton` is introduced to indicate whether the autofill form button should be shown in the cell. The flow of this data happens the following way:
For datatype A, `ExpandedManualFillCoordinator` -> `ACoordinator` -> `AMediator` -> `AItemType` -> `AItemCell`.
Then `AItemCell` is responsible for the showing/hiding the button.

Screencast: http://screencast/cast/NTMwODY1OTk4NTgwOTQwOHxjZmQzOTM0YS1iOA
Bug: 338208859
Change-Id: Ib6dad7389979823f227743c5dd259ccd096de739
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5692078
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Reviewed-by: Noémie St-Onge <noemies@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1329713}