[Enterprise] Include if more than the user is attending an event
The calendar card on the NTP will prioritize showing more information
for an event if it has an attendee other than the user that is not a
resource and has not declined.

In order to support this, I've added a has_other_attendee field to
the CalendarEvent object. I thought about passing through all the
attendees, but that's more information than necessary.

Bug: 351921215
Change-Id: If67f11b9102976e1247311eac9d035e8e790f2f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5718849
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Commit-Queue: Riley Tatum <rtatum@google.com>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329723}