lorenz: Remove obsolete param and add typing
Now that the bot no longer uses `-j`, and we have migrated away from
calling ninja directly, the `-j` param no longer makes sense, as we are
either calling autoninja locally or siso on the bot, neither of which
uses a `-j` param.

Also add an OptionsNamespace similar to `generate_profile.py` as this
allows the IDE to easily understand the allowed args and their various
types. Adding these types also exposed flaws in the previous types, for
example, previously `prefixes` was typed as `Tuple[str]`, which actually
meant a tuple of size exactly one and containing a single str element.
This was not caught by the type checker in VS Code since the only param
passed was `tuple(args.prefixes)` and since the type checker was unable
to discern what `args.prefixes` contains, it passed type checking. Now
that args.prefixes is defined to be `List[str]`, that is, a list of one
or more str elements, the type checker correctly discerns that the
`Tuple[str]` type is incorrect, it should instead be `Tuple[str, ...]`,
see [0]. Adding type annotations also made it clear that
`args.build_output_dir` is passed as a string argument but is then
converted to a pathlib.Path path.

[0] https://docs.python.org/3/library/typing.html#annotating-tuples

Bug: 348423879
Change-Id: I47628dd3cb294782318a07be86298ad8a9f03ff6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5722734
Auto-Submit: Peter Wen <wnwen@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329802}