[Omnibox] Fix answer cache for suggestions with answer_template
This cache allows psuggest search suggestions to be augmented with an
associated answer via a prefetch. The prefetch is only performed if
there is a cached entry corresponding to the answer; currently this
entry is only updated for suggestions with the old SuggestionAnswer
structure. This change makes the cache agnostic to the answer
representation; it caches the full text for any suggestion in the top 2
rows with an answer type that's not UNSPECIFIED.

Bug: 327497146
Change-Id: Ie043dfa00fa9966df5b4d711871460786d95deb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5722512
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Jennifer Serrano <jennserrano@google.com>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329835}