Stop playback when backing to newtab page.
onLoadStarted isn't always called, for example in the situation above. As an extra precaution against continuing playback when we shouldn't, I've added a method that will stop the playback if the url of a playing tab changes.

Bug: 328672981
Change-Id: I5491b7dcc185589f9689be235068c11c754afe74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5717519
Reviewed-by: Ian Wells <iwells@chromium.org>
Commit-Queue: Basia Zimirska <basiaz@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1329838}