birch: Only fetch weather info in the mornings
The weather chip is only supposed to show in the mornings. Only do a
server fetch in the mornings, since we won't show weather the rest of
the day anyway. This will reduce load on the weather server.

Remove a hack that was showing weather in the afternoons. If you need
to see weather in your local afternoon, run your code with the switch
--birch-is-morning. This CL also adds switch --birch-is-evening to
force it not to be morning (and hence not to fetch weather).

Change some BirchModel unit tests to not assert a particular order for
BirchItems. This means they don't have to be modified if ranking
changes.

Bug: 353762809
Test: added to ash_unittests
Change-Id: I8e3c3edd8e58fcf456db4771d810ec55e3298156
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5722214
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329849}