[pip] [reland] Show capture media picker in pip if it's focused
If `getDisplayMedia()` is called on a pip window's opener, but
the pip window currently has the focus, then show the UI in the
pip window instead.  It's common to open the stream in the opener
so that its lifetime isn't tied to the pip window, but it can be
confusing if the UI shows up in the opener while the user is
interacting with the pip window.

The reland uses widget activation in place of RWHV::IsFocused,
since the latter returns true basically all the time on mac.
While this might indicate an issue with IsFocused that should
be fixed, IsActive is what's used by the frame to determine
how the window borders are drawn anyway.  This way, they're
guaranteed to be in sync.

This CL also replaces the unit test with a browser test that
changes focus, to catch any per-platform behavior.  The unit
test was not maintainable, unfortunately, because the switch
to check Widget activation requires views rather than just
content//.  The browser test captures all of what the unit
test did, albeit with all the drawbacks of a browser test.

PS1 is the original CL, for comparison.

Original CL:
https://chromium-review.googlesource.com/c/chromium/src/+/5536096

Bug: 328673345, 40618765

 Conflicts:
	content/test/test_render_view_host.h

Change-Id: Ice72f6515b60ceec40a137ef8a749c108a2bbd84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5661599
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329858}