[Read Aloud] Reset to unsupported boundary state after a voice change.
When we change voices, we no longer know if word boundaries are
supported for the given voice, so to be safe, reset the word
boundary state.

After we receive a word boundary event, we can continue to use
word boundaries with the current voice.

This has the side effect of word highlighting not working immediately
after a voice change. Additionally, if a voice changes and we pause
immediately after before another word boundary has been retrieved,
we won't be able to resume on the current word. However, these are
preferable to incorrectly showing a word boundary for an
unsupported voice.

Fixed: 353566103
Change-Id: Ic516ef64f3a119a890cbae199a686e6bbc219d62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5715330
Reviewed-by: Eitan Goldberger <eitang@google.com>
Commit-Queue: Lauren Winston <lwinston@google.com>
Cr-Commit-Position: refs/heads/main@{#1329886}