Default initialize OpenXR views to prevent invalid xrEndFrame calls
Some users have reported errors (particularly/only when using depth...)
with sessions terminating early because poses were unable to be found.
This lead to to seeing `XR_ERROR_POSE_INVALID` in the logs, because we
had a non-unit quaternion (per the spec). This default-initializes the
views so that these early frames don't fail if the views are not set,
though the default FOV values will likely make content not visible, and
there is a risk that this could hide a real error if views are never
located, that risk seems small compared to bailing out of sessions
simply because they are still getting things spun up.

Fixed: 353924002
Change-Id: I0961c4b40fa5f3d5558d1f7ff49b5021c9e8227d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5722936
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329892}