[Traffic Annotations] Handle repeated policy values
When printing to the .tsv file after auditing, the auditor parses the
policy values in the traffic annotation proto, passing values into
text_format.PrintField. If the value is a repeated field, this gets
passed as a List[value type], but PrintField requires only a single
value.

Add a conditional for this case, and loop through the values of the
field in the repeated case.

Bug: 354004347
Change-Id: If1248703a8b005e7f76587ae139ebd2c6f5623b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5722932
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Commit-Queue: Chris Mullins <crmullins@chromium.org>
Reviewed-by: Aiden Chiavatti <chiav@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1329893}