[lensoverlay] Fix tooltip showing on more info menu
The most robust way to do this is change the logic for handling when to
show the exit tooltip. Instead of having it controlled by multiple
elements, all the logic is based on hovering the background scrim.

Also added testing for the cursor tooltip.

Bug: 347732589
Change-Id: I0ae544c47d726e63fde3518f4ddc1e6df1b52269
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5723181
Commit-Queue: Duncan Mercer <mercerd@google.com>
Reviewed-by: Bryan Nguyen <nguyenbryan@google.com>
Cr-Commit-Position: refs/heads/main@{#1329901}