[Read Aloud] Improve logic for choosing a locale for en, es, & pt
By choosing an arbitrary locale, we were enabling and downloading
locales for these languages even if a different locale for that language
was already enabled or downloaded. With this change, we first check if
there's a relevant locale already enabled or at least downloaded before
choosing the arbitrary locale.

Fixed: 351885607
Change-Id: I8fea8760a56ef03d2742fec7616dc336b52390a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5719531
Reviewed-by: Eitan Goldberger <eitang@google.com>
Commit-Queue: Kristi Saney <kristislee@google.com>
Cr-Commit-Position: refs/heads/main@{#1329917}